+++ CHANGELOG.md
@@ -32,6 +32,8 @@ All notable changes to this project will be documented in this file.
 
 - Added a function to DPR estimates to group Cornwall and Isles of Scilly.
 
+- Created job role estimates data validation columns in Polars and added validation and test validation scripts for the Estimates job. 
+
 ### Changed
 - Removed the PySpark version of IND CQC Clean and Validation jobs.